Проведен крупный рефакторинг всего сервиса Сервис разделен на модули:   parent - выступает в роли центра для всех зависимостей   fw - конфигурации   app - api/impl - бизнес логика сервиса   domain - модели и сущности   adapter - адаптеры для рест запросов, базы данных и пр.
