+++ Dockerfile
@@ -1,0 +1,5 @@
+FROM adoptopenjdk/openjdk11:alpine-jre
+EXPOSE 8080
+ARG JAR_FILE=target/*.jar
+COPY ${JAR_FILE} app.jar
+CMD java -jar app.jar
