+++ app/app-api/src/main/java/com/devilpanda/nota/tabula/app/api/IncorrectEmailException.java
@@ -1,0 +1,19 @@
+package com.devilpanda.nota.tabula.app.api;
+
+public class IncorrectEmailException extends RuntimeException {
+    public IncorrectEmailException() {
+        super();
+    }
+
+    public IncorrectEmailException(String message) {
+        super(message);
+    }
+
+    public IncorrectEmailException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public IncorrectEmailException(Throwable cause) {
+        super(cause);
+    }
+}
