+++ app/app-api/src/main/java/com/devilpanda/nota/tabula/app/api/UserAlreadyExistException.java
@@ -1,0 +1,20 @@
+package com.devilpanda.nota.tabula.app.api;
+
+public class UserAlreadyExistException extends RuntimeException {
+
+    public UserAlreadyExistException() {
+        super();
+    }
+
+    public UserAlreadyExistException(String message) {
+        super(message);
+    }
+
+    public UserAlreadyExistException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public UserAlreadyExistException(Throwable cause) {
+        super(cause);
+    }
+}
