+++ docker-compose.yml
@@ -1,0 +1,15 @@
+version: '3'
+
+services:
+  app:
+    image: 'nota-tabula-back:latest'
+    network_mode: host
+    build:
+      context: .
+    container_name: app
+    depends_on:
+      - db
+    environment:
+      - SPRING_DATASOURCE_URL=jdbc:postgresql://localhost:5432/nota_tabula_db
+      - SPRING_DATASOURCE_USERNAME=sa
+      - SPRING_DATASOURCE_PASSWORD=password
