+++ domain/src/main/java/com/devilpanda/nota/tabula/domain/Role.java
@@ -1,0 +1,22 @@
+package com.devilpanda.nota.tabula.domain;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.annotations.GenericGenerator;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "ROLE")
+@Getter
+@Setter
+public class Role {
+    @Id
+    @GeneratedValue(generator = "increment")
+    @GenericGenerator(name = "increment", strategy = "increment")
+    private Long id;
+
+    @Enumerated(EnumType.STRING)
+    private RoleName name;
+
+}
