+++ domain/src/main/java/com/devilpanda/nota/tabula/domain/RoleName.java
@@ -1,0 +1,5 @@
+package com.devilpanda.nota.tabula.domain;
+
+public enum RoleName {
+    ROLE_USER
+}
