+++ fw/src/main/java/com/devilpanda/nota/tabula/fw/security/dto/LoginFormDto.java
@@ -1,0 +1,9 @@
+package com.devilpanda.nota.tabula.fw.security.dto;
+
+import lombok.Data;
+
+@Data
+public class LoginFormDto {
+    private String email;
+    private String password;
+}
