+++ fw/src/main/java/com/devilpanda/nota/tabula/fw/security/dto/RegisterFormDto.java
@@ -1,0 +1,10 @@
+package com.devilpanda.nota.tabula.fw.security.dto;
+
+import lombok.Data;
+
+@Data
+public class RegisterFormDto {
+    private String login;
+    private String email;
+    private String password;
+}
