+++ fw/src/main/java/com/devilpanda/nota/tabula/fw/security/jwt/JwtFilter.java
@@ -1,0 +1,53 @@
+package com.devilpanda.nota.tabula.fw.security.jwt;
+
+import com.devilpanda.nota.tabula.fw.security.UserDetailService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.GenericFilterBean;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
+import javax.servlet.http.HttpServletRequest;
+import java.io.IOException;
+
+@Component
+public class JwtFilter extends GenericFilterBean {
+
+    private static final String AUTHORIZATION = "Authorization";
+
+    @Autowired
+    private JwtProvider jwtProvider;
+    @Autowired
+    private UserDetailService userDetailService;
+
+    @Override
+    public void doFilter(ServletRequest servletRequest, ServletResponse servletResponse, FilterChain filterChain) throws IOException, ServletException {
+        try {
+            String token = getJwtToken((HttpServletRequest) servletRequest);
+            logger.info("Filter working... token: " + token);
+
+            if (token != null && jwtProvider.validateToken(token)) {
+                String userName = jwtProvider.getUserNameFromToken(token);
+                UserDetails userDetails = userDetailService.loadUserByUsername(userName);
+                UsernamePasswordAuthenticationToken auth
+                        = new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
+                SecurityContextHolder.getContext().setAuthentication(auth);
+            }
+        } catch (Exception e) {
+            logger.error("Can't set user auth", e);
+        }
+        filterChain.doFilter(servletRequest, servletResponse);
+    }
+
+    private String getJwtToken(HttpServletRequest request) {
+        String authHeader = request.getHeader(AUTHORIZATION);
+        if (authHeader != null && authHeader.startsWith("Bearer"))
+            return authHeader.replace("Bearer ", "");
+        return null;
+    }
+}
