+++ fw/src/main/java/com/devilpanda/nota/tabula/fw/security/jwt/JwtResponse.java
@@ -1,0 +1,15 @@
+package com.devilpanda.nota.tabula.fw.security.jwt;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class JwtResponse {
+    private String accessToken;
+    private String type = "Bearer";
+
+    public JwtResponse(String accessToken) {
+        this.accessToken = accessToken;
+    }
+}
