+++ fw/src/main/resources/db/migration/V2__create_users.sql
