+++ fw/src/test/java/com/devilpanda/nota/tabula/fw/NotaTabulaApplicationTests.java
@@ -1,0 +1,13 @@
+package com.devilpanda.nota.tabula.fw;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NotaTabulaApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
