--- src/main/java/com/devilpanda/nota_tabula/NotaTabulaApplication.java
@@ -1,13 +1,0 @@
-package com.devilpanda.nota_tabula;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NotaTabulaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(NotaTabulaApplication.class, args);
-	}
-
-}
