--- src/main/java/com/devilpanda/nota_tabula/controller/TestController.java
@@ -1,15 +1,0 @@
-package com.devilpanda.nota_tabula.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/test")
-public class TestController {
-
-    @GetMapping("")
-    public String test() {
-        return "Access granted!";
-    }
-}
