--- src/main/java/com/devilpanda/nota_tabula/errors/UserAlreadyExistException.java
@@ -1,20 +1,0 @@
-package com.devilpanda.nota_tabula.errors;
-
-public class UserAlreadyExistException extends RuntimeException {
-
-    public UserAlreadyExistException() {
-        super();
-    }
-
-    public UserAlreadyExistException(String message) {
-        super(message);
-    }
-
-    public UserAlreadyExistException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public UserAlreadyExistException(Throwable cause) {
-        super(cause);
-    }
-}
