--- src/main/java/com/devilpanda/nota_tabula/model/Role.java
@@ -1,22 +1,0 @@
-package com.devilpanda.nota_tabula.model;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.GenericGenerator;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "ROLE")
-@Getter
-@Setter
-public class Role {
-    @Id
-    @GeneratedValue(generator = "increment")
-    @GenericGenerator(name = "increment", strategy = "increment")
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private RoleName name;
-
-}
