--- src/main/java/com/devilpanda/nota_tabula/model/RoleName.java
@@ -1,5 +1,0 @@
-package com.devilpanda.nota_tabula.model;
-
-public enum RoleName {
-    ROLE_USER
-}
