--- src/main/java/com/devilpanda/nota_tabula/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.devilpanda.nota_tabula.repository;
-
-import com.devilpanda.nota_tabula.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByLogin(String login);
-
-    Optional<User> findByEmail(String email);
-
-}
