--- src/main/java/com/devilpanda/nota_tabula/security/UserPrincipal.java
@@ -1,67 +1,0 @@
-package com.devilpanda.nota_tabula.security;
-
-import com.devilpanda.nota_tabula.model.User;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import lombok.AllArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@AllArgsConstructor
-public class UserPrincipal implements UserDetails {
-    private final String email;
-
-    @JsonIgnore
-    private final String password;
-
-    private final Collection<? extends GrantedAuthority> authorities;
-
-    public static UserPrincipal build(User user) {
-        List<GrantedAuthority> authorities = user.getRoles().stream().map(role ->
-                new SimpleGrantedAuthority(role.getName().name())
-        ).collect(Collectors.toList());
-        return new UserPrincipal(
-                user.getEmail(),
-                user.getPassword(),
-                authorities
-        );
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return authorities;
-    }
-
-    @Override
-    public String getPassword() {
-        return password;
-    }
-
-    public String getUsername() {
-        return email;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
