--- src/main/java/com/devilpanda/nota_tabula/security/dto/LoginFormDto.java
@@ -1,9 +1,0 @@
-package com.devilpanda.nota_tabula.security.dto;
-
-import lombok.Data;
-
-@Data
-public class LoginFormDto {
-    private String email;
-    private String password;
-}
