--- src/main/java/com/devilpanda/nota_tabula/security/jwt/JwtResponse.java
@@ -1,15 +1,0 @@
-package com.devilpanda.nota_tabula.security.jwt;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class JwtResponse {
-    private String accessToken;
-    private String type = "Bearer";
-
-    public JwtResponse(String accessToken) {
-        this.accessToken = accessToken;
-    }
-}
