--- src/main/resources/application.properties
@@ -1,5 +1,0 @@
-spring.datasource.driver-class-name=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/nota_tabula_db
-spring.datasource.username=sa
-spring.datasource.password=password
-spring.jpa.show-sql=true
