--- src/test/java/com/devilpanda/nota_tabula/NotaTabulaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.devilpanda.nota_tabula;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NotaTabulaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
